Refactoring sov-module-api
